Revert "Add Tests and Code for XML21 support"
This reverts commit 360874d4ecff9379efd3c19edf1d0a495db5fe40.
