+++ composer.json
@@ -31,9 +31,7 @@
         "guzzlehttp/guzzle": "^6.3",
         "vlucas/valitron": "^1.4",
         "ext-SimpleXML": "*",
-        "ext-json": "*",
-        "ext-dom": "*",
-        "findologic/xml-response-schema": "^2.1"
+        "ext-json": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "^5 || ^6.5 || ^7.0",
