+++ src/FINDOLOGIC/Api/Responses/Xml20/Properties/LandingPage.php
@@ -5,9 +5,6 @@ namespace FINDOLOGIC\Api\Responses\Xml20\Properties;
 use FINDOLOGIC\Api\Helpers\ResponseHelper;
 use SimpleXMLElement;
 
-/**
- * @deprecated Use XML 2.1 instead. This class will be removed with version v1.0.0-rc.1.
- */
 class LandingPage
 {
     /** @var string $link */
