+++ src/FINDOLOGIC/Api/Responses/Xml20/Xml20Response.php
@@ -15,8 +15,6 @@ use SimpleXMLElement;
 
 /**
  * Is used for search and navigation requests with XML response only!
- * 
- * @deprecated Use XML 2.1 instead. This class will be removed with version v1.0.0-rc.1.
  */
 class Xml20Response extends Response
 {
