--- src/FINDOLOGIC/Api/Responses/Xml21/Properties/LandingPage.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace FINDOLOGIC\Api\Responses\Xml21\Properties;
-
-use FINDOLOGIC\Api\Helpers\ResponseHelper;
-use SimpleXMLElement;
-
-class LandingPage
-{
-    /** @var string $link */
-    private $link;
-
-    /**
-     * LandingPage constructor.
-     * @param SimpleXMLElement $response
-     */
-    public function __construct($response)
-    {
-        $this->link = ResponseHelper::getStringProperty($response, 'link');
-    }
-
-    /**
-     * @return string
-     */
-    public function getLink()
-    {
-        return $this->link;
-    }
-}
