--- tests/FINDOLOGIC/Api/Mockdata/Xml21/demoResponseWithoutAllowOverride.xml
@@ -1,169 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<searchResult>
-    <servers>
-        <frontend>martell.frontend.findologic.com</frontend>
-        <backend>hydra.backend.findologic.com</backend>
-    </servers>
-    <query>
-        <limit first="0" count="24"/>
-        <queryString>ps3</queryString>
-        <originalQuery>original query</originalQuery>
-        <didYouMeanQuery>ps4</didYouMeanQuery>
-    </query>
-    <landingPage link="https://www.landingpage.io/agb/"/>
-    <promotion link="https://promotion.com/" image="https://promotion.com/promotion.png"/>
-    <results>
-        <count>1808</count>
-    </results>
-    <products>
-        <product id="019111105-37900" relevance="1.3862943649292">
-            <properties>
-                <property name="ordernumber">019111105-37900</property>
-            </properties>
-        </product>
-        <product id="029214085-37860" relevance="1.3862943649292"/>
-    </products>
-    <filters>
-        <main>
-            <filter>
-                <name>price</name>
-                <display>Preis</display>
-                <select>single</select>
-                <type>range-slider</type>
-                <attributes>
-                    <selectedRange>
-                        <min>0.39</min>
-                        <max>2239.1</max>
-                    </selectedRange>
-                    <totalRange>
-                        <min>0.39</min>
-                        <max>2239.1</max>
-                    </totalRange>
-                    <stepSize>0.1</stepSize>
-                    <unit>€</unit>
-                </attributes>
-                <items>
-                    <item>
-                        <name>0.39 - 13.4</name>
-                        <weight>0.51743012666702</weight>
-                        <parameters>
-                            <min>0.39</min>
-                            <max>13.4</max>
-                        </parameters>
-                    </item>
-                    <item>
-                        <name>13.45 - 25.99</name>
-                        <weight>0.50098878145218</weight>
-                        <parameters>
-                            <min>13.45</min>
-                            <max>25.99</max>
-                        </parameters>
-                    </item>
-                    <item>
-                        <name>26 - 40.3</name>
-                        <weight>0.3976277410984</weight>
-                        <parameters>
-                            <min>26</min>
-                            <max>40.3</max>
-                        </parameters>
-                    </item>
-                </items>
-            </filter>
-            <filter>
-                <name>Farbe</name>
-                <display>Farbe</display>
-                <select>multiselect</select>
-                <selectedItems>1</selectedItems>
-                <type>color</type>
-                <items>
-                    <item>
-                        <name>beige</name>
-                        <weight>0.10730088502169</weight>
-                        <image>https://blubbergurken.io/farbfilter/beige.gif</image>
-                        <color>#F5F5DC</color>
-                    </item>
-                    <item>
-                        <name>blau</name>
-                        <weight>0.3296460211277</weight>
-                        <image>https://blubbergurken.io/farbfilter/blau.gif</image>
-                        <color>#3c6380</color>
-                    </item>
-                    <item>
-                        <name>braun</name>
-                        <weight>0.90265488624573</weight>
-                        <image>https://blubbergurken.io/farbfilter/braun.gif</image>
-                        <color>#94651e</color>
-                    </item>
-                </items>
-            </filter>
-        </main>
-        <other>
-            <filter>
-                <name>Material</name>
-                <display>Material</display>
-                <select>multiple</select>
-                <type>select</type>
-                <items>
-                    <item>
-                        <name>Hartgepäck</name>
-                        <weight>0.038716815412045</weight>
-                        <frequency>35</frequency>
-                    </item>
-                    <item>
-                        <name>Leder</name>
-                        <weight>0.63053095340729</weight>
-                        <frequency>1238</frequency>
-                    </item>
-                    <item>
-                        <name>Nylon</name>
-                        <weight>0.12168141454458</weight>
-                        <frequency>110</frequency>
-                    </item>
-                </items>
-            </filter>
-            <filter>
-                <name>vendor</name>
-                <display>Hersteller</display>
-                <select>multiple</select>
-                <type>select</type>
-                <items>
-                    <item>
-                        <name>Bodenschatz</name>
-                        <weight>0.0022123893722892</weight>
-                        <frequency>2</frequency>
-                    </item>
-                    <item>
-                        <name>Braun Büffel</name>
-                        <weight>0.08517698943615</weight>
-                        <frequency>77</frequency>
-                    </item>
-                    <item>
-                        <name>Camel Active</name>
-                        <weight>0.13495574891567</weight>
-                        <frequency>122</frequency>
-                    </item>
-                </items>
-            </filter>
-            <filter>
-                <name>cat</name>
-                <display>Kategorie</display>
-                <select>single</select>
-                <type>select</type>
-                <items>
-                    <item>
-                        <name>Buch</name>
-                        <weight>0.25156819820404</weight>
-                        <frequency>5</frequency>
-                        <items>
-                            <item>
-                                <name>Beste Bücher</name>
-                                <weight>0.33799207210541</weight>
-                                <frequency>5</frequency>
-                            </item>
-                        </items>
-                    </item>
-                </items>
-            </filter>
-        </other>
-    </filters>
-</searchResult>
