+++ tests/FINDOLOGIC/Api/Mockdata/demoResponseWithOriginalQueryType.xml
