+++ tests/FINDOLOGIC/Api/Mockdata/demoResponseWithoutPromotion.xml
