+++ tests/FINDOLOGIC/Api/Tests/Requests/SearchNavigation/SearchNavigationRequestTest.php
@@ -35,7 +35,7 @@ class SearchNavigationRequestTest extends TestBase
             ->setServiceId('ABCDABCDABCDABCDABCDABCDABCDABCD')
             ->setHttpClient($this->httpClientMock);
 
-        $this->rawMockResponse = $this->getMockResponse('Xml20/demoResponse.xml');
+        $this->rawMockResponse = $this->getMockResponse('demoResponse.xml');
     }
 
     public function requiredParamsProvider()
